add the Public Setting
In the setting-public.php file, a user's session is initiated to check if they are logged in. If not, they are redirected to the login page. Once authenticated, a connection is established to the database. On the frontend, there are input fields for the user to input or edit the public settings of the website, including the site name, tag line, and domain name. When the form is submitted, the data is saved to the database. There are also built-in alert mechanisms that display success or error messages to the user. These messages will automatically disappear after a few seconds. The DataTables jQuery plugin is also initialized for any data tables present on the page.